Move color types to math/Color.h
